Introduced gfx_dep_sigrf and gfx_dep_sigraf

+++ include/groufix/def.h
@@ -91,7 +91,7 @@
  * Define an operator| for enum types in C++ mode.
  */
 #if defined (__cplusplus)
-	#define GFX_BIT_FIELD(T) inline T operator|(T a, T b) { return (T)((int)a | (int)b); }
+	#define GFX_BIT_FIELD(T) inline T operator|(T a, T b) { return (T)((long)a | (long)b); }
 #else
 	#define GFX_BIT_FIELD(T)
 #endif
